services/api_service/service.go:standalone apiService
Created api_service (apiService) as a service to decouple
main.go from the api service implementation.

This allows to run the api service without utilizing main.
It also allows for separte tests of the api service itself,
as well as tests to utilize the package separate.

* Moved api service implementation from main to service *
